+++ pkg/util/validation/validate_test.go
@@ -94,7 +94,7 @@ func TestValidateLabels(t *testing.T) {
 		{
 			map[model.LabelName]model.LabelValue{model.MetricNameLabel: "badLabelValue", "much_shorter_name": "test_value_please_ignore_no_really_nothing_to_see_here"},
 			false,
-			httpgrpc.Errorf(http.StatusBadRequest, errLabelValueTooLong, "test_value_please_ignore_no_really_nothing_to_see_here", `badLabelValue{much_shorter_name="test_value_please_ignore_no_really_nothing_to_see_here"}`),
+			httpgrpc.Errorf(http.StatusBadRequest, errLabelValueTooLong, `badLabelValue{much_shorter_name="test_value_please_ignore_no_really_nothing_to_see_here"}`, "test_value_please_ignore_no_really_nothing_to_see_here"),
 		},
 		{
 			map[model.LabelName]model.LabelValue{model.MetricNameLabel: "foo", "bar": "baz", "blip": "blop"},